requirements.txt
anyio==3.7.0
boltons==23.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
conda==23.3.1
conda-package-handling==2.0.2
conda_package_streaming==0.8.0
cryptography==41.0.0
exceptiongroup==1.1.1
fastapi==0.95.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jsonpatch==1.32
jsonpointer==2.0
libmambapy==1.4.2
mamba==1.4.2
packaging==23.1
pip==23.1.2
pluggy==1.0.0
pycosat==0.6.4
pycparser==2.21
pydantic==1.10.8
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
setuptools==67.7.2
sniffio==1.3.0
starlette==0.27.0
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.6.2
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wheel==0.40.0
zstandard==0.19.0
httpx~=0.24.1